setup/requirements.txt
Flask==3.1.1
langgraph==0.6.3
langchain==0.3.27
langchain_community==0.3.27
langchain_ollama==0.3.6
python-dotenv==1.1.1
pdfplumber==0.11.7
unstructured==0.18.11
langchain_chroma==0.2.5
docx2txt==0.9
python-docx==1.2.0
PyPDF2==3.0.1
pywin32==311
langgraph-checkpoint-sqlite==2.0.11
rank-bm25==0.2.2
nltk==3.9.1
quart==0.20.0
hypercorn==0.17.3
langchain-milvus==0.2.1
protobuf==5.27.2
aiosqlite==0.21.0
scikit-learn==1.7.1